requirements.txt
podcastfy
python-dotenv
pyyaml 